Revert "Don't delete obsolete tank templates, and permit changing templates without resetting all factories."
This reverts commit 9d7caf170753e1dd5e8d9be9ad9ee86cbc36bacb.

This is too drastic a change for the game. I didn't think it'd effect gameplay, but it does touch the nostalgia a bit too much. Will add a commit to simply permit deleted templates to continue producing until aborted. 2.3.x let you keep the template, but that's a bit hard to do.